feat : response 고도화 refactoring - package refactoring - exception handling refactoring
