+++ src/main/java/com/ureca/ocean/jjh/client/dto/UserDto.java
@@ -16,6 +16,4 @@ public class UserDto {
 	private String name;
 	private String email;
 	private String password;
-	
-	private Map<Integer, String> roles;
 }
