+++ src/main/java/com/ureca/ocean/jjh/common/exception/ErrorCode.java
@@ -6,9 +6,14 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public enum ErrorCode {
-    //INTERNAL SERVER ERROR
-    INTERNAL_SERVER_ERROR(20000, "INTERNAL_SERVER_ERROR",  "서버 내부 오류가 발생했습니다."),
-    BAD_REQUEST(00401,"BAD_REQUEST","BODY가 비어있습니다.");
+    //auth_backend : 10000
+    LOGIN_FAIL(10000,"LOGIN_FAIL","LOGIN에 실패했습니다."),
+    //user_backend : 20000
+    NOT_FOUND_USER(20001,"NOT_FOUND_USER","해당 사용자가 없습니다."),
+    USER_ALREADY_EXIST(20002,"USER_ALREADY_EXIST","해당 사용자가 이미 있습니다.");
+
+    //map_backend : 30000
+
     private final int code;
     private final String name;
     private final String message;
