+++ src/main/java/com/ureca/ocean/jjh/exception/AuthException.java
@@ -1,0 +1,11 @@
+package com.ureca.ocean.jjh.exception;
+
+import com.ureca.ocean.jjh.common.exception.ErrorCode;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Getter
+public class AuthException extends RuntimeException{
+    private final ErrorCode errorCode;
+}
